Add files via upload
Calculation of the polarizability using a small k-mesh around the Fermi level (instead of integrating all the 1BZ). The hamiltonian is diagonalized only once at each point in the k-mesh, so the calculation of the polarizability now takes only a few minutes (notes on preparation).  Because of the smallness of the k-mesh, it is not necessary to use periodic boundary conditions in this case.